listing-sniper/requirements.txt
# Core async
aiohttp>=3.9.0
websockets>=12.0

# Solana
solders>=0.21.0
solana>=0.34.0

# Database
asyncpg>=0.29.0
redis>=5.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Monitoring
prometheus-client>=0.20.0

# Telegram
aiohttp>=3.9.0

# CEX data (optional, for backtesting)
ccxt>=4.2.0

# Type checking & linting
mypy>=1.8.0
ruff>=0.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aioresponses>=0.7.6
